requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1

langchain==0.3.7
langchain-core==0.3.18
langchain-community==0.3.7
langserve==0.3.0

chromadb==0.5.23
tiktoken==0.8.0

requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

langchain-openai==0.2.9
